Fix cc error handling and more style tweaks
