+++ src/source/raw.ts
@@ -2,15 +2,13 @@ import chardet from 'chardet';
 import request from 'request';
 import { userAgent, cookie } from '../config';
 import { TextDecoder } from 'util';
+import { Source } from '../utils';
 
-export default async function fetchRaw(
-  name: string,
-  url: string
-): Promise<string> {
+export default async function fetchRaw(source: Source): Promise<string> {
   return new Promise((resolve, reject) => {
     request(
       {
-        uri: url,
+        uri: source.url,
         headers: {
           Cookie: cookie,
           'User-Agent': userAgent,
@@ -24,14 +22,14 @@ export default async function fetchRaw(
             resolve(new TextDecoder(encoding).decode(body));
           } catch (e) {
             console.log(
-              `[${name}]: failed to decode response. detected encoding:${encoding} err:${e}`
+              `[${source.name}]: failed to decode response. detected encoding:${encoding} err:${e}`
             );
             resolve(body.toString());
           }
         } else if (!body.length) {
-          reject(`[${name}]: empty response`);
+          reject('empty response');
         } else if (res.statusCode !== 404) {
-          reject(`[${name}]: ERROR ${res.statusCode} err:${err}`);
+          reject(`ERROR ${res.statusCode} err:${err}`);
         }
       }
     );
